Reduce Call Answer Delay
This patch sets a boolean acceptingIncomingCall to true on
acceptCall.  In setAudioMode, we set the mode as IN_CALL if the call is
ringing and acceptingIncomingCall is true.  This initializes the
microphone sooner, and should reduce the microphone delay.

Not sure if this is the best way to do this, but this reduces call
delay to under a second for me.  I thought I might as well upload it so
others could have a look at it.

Change-Id: Ied21acb81c8447fc9949cecf861c0be4ca67ebb0