Fix battery icon flashing quickly on percent change
If you use the "show battery % on statusbar" option, when the percent
changes, the "traditional" battery indicator reappears for a fraction
of a second. Probably this is most noticeable on lower end phones.
This commit prevents that other icon from appearing when the battery
percent on statusbar option is enabled.

Change-Id: Icd593754c5879b84139ffbc85c30f7ecb5e258a4