Option to disable lockscreen if security screen lock active (1/3)
[This patch consists of three commits. This one, one for Settings and one for CMParts.]

Adds an option to CMParts that allows to turn off lockscreen if
a security screen lock is enabled.

Default value for new users is "disabled", so the lockscreen is always displayed before
the security screen.

If you update from an earlier build of CM, this option will be automatically enabled if pattern lock is
active and otherwise disabled.

Change-Id: I03a9b3815329f4541debbfd4255e64d08d6a3b9d