Add headset icon toggle (Framework)
What:
Adds a setting that enables/disables the headset icon from displaying on the
status bar when a wired headset is plugged in.

Why:
Many users have requested this feature because the icon wastes space on their
status bar and is somewhat unnecessary. Also, other UI's for android do not
include a headset icon (Sense, for example), so users may be used to not seeing
the icon and prefer it that way. Users with this preference can now disable
the icon from ever appearing.

Change-Id: I0b066af15e5d2534a29cc7a6b5bea222c30ea1e0